requirements.txt
dbt-sqlserver
azure-identity
azure-keyvault-secrets
azure-appconfiguration
